Crate window_manager

This crate acts as a manager of a list of windows. It defines a WindowManager structure and an instance of it.

A window manager holds a set of WindowInner objects, including an active window, a list of shown windows and a list of hidden windows. The hidden windows are totally overlapped by others.

A window manager owns a bottom framebuffer and a top framebuffer. The bottom is the background of the desktop and the top framebuffer contains a floating window border and a mouse arrow. A window manager also contains a final framebuffer which is mapped to the screen. In refreshing an area, the manager will render all the framebuffers to the final one in order: bottom -> hide list -> showlist -> active -> top.

The window manager provides methods to update within some bounding boxes rather than the whole screen for better performance.

Structs

WindowManager
Window manager structure which maintains a list of windows and a mouse.

Statics

WINDOW_MANAGER
The instance of the default window manager

Functions

init
Initialize the window manager. It returns (keyboard_producer, mouse_producer) for the I/O devices.